added so that you can access the content from the templates

+++ cms/assets/js/cms.js
@@ -327,6 +327,14 @@
 })(Date);
 
 
+function slugify(text) {
+  return text.toString().toLowerCase().trim()
+    .replace(/\s+/g, '-')
+    .replace(/&/g, '-and-')
+    .replace(/[^\w\-]+/g, '')
+    .replace(/\-\-+/g, '-');
+}
+
 function CMS() {
   var saveButtons = [].slice.call(document.querySelectorAll(".js-save-doc"));
   var saveContentTypeButtons = [].slice.call(document.querySelectorAll(".js-save-content-type"));
@@ -337,8 +345,18 @@ function CMS() {
   var deleteDocButton = document.querySelector(".js-delete-doc");
   var deleteContentTypeButton = document.querySelector(".js-delete-content-type");
 
+  var slugger = document.querySelector('.slugger');
+  if (slugger) {
+    var slugField = document.querySelector('input[name="_name"]');
+    slugField.addEventListener('input', function() {
+      slugger.value = slugify(slugField.value);
+    });
+  }
+
   if (deleteDocButton) {
     deleteDocButton.addEventListener('click', function() {
+      if (!confirm("Are you sure?")) return;
+      
       var url = '/admin/doc/' + doc._id;
       var method = 'DELETE';
 
@@ -358,6 +376,8 @@ function CMS() {
 
   if (deleteContentTypeButton) {
     deleteContentTypeButton.addEventListener('click', function() {
+      if (!confirm("Are you sure?")) return;
+
       var url = '/admin/content-type/' + contentType._id;
       var method = 'DELETE';
